fix: purge static pages on deploy by tagging them "short"
Every static, GTFS-derived response was tagged Cache-Tag: long, but the
post-deploy cache drop in cloudbuild.yaml only purges tags=short. A data
refresh that added a route therefore left /routes and friends serving a
pre-import copy from Cloudflare for up to the full 30-day s-maxage:
route А67 was missing from /routes for ~28h after the 2026-08-01 import
even though the origin had it.

Switch every long-tagged response to "short" so the deploy purge reaches
it. TTLs are unchanged; with deploy-time purging, s-maxage no longer
governs freshness. Renames the now-misleading longCacheAgeSeconds local
to cacheAgeSeconds.

Also emit the Latin route id in the /routes listing hrefs via the
existing routeNameToUrlFriendly helper, so links land on /route/A67
directly instead of taking a 301 from the Cyrillic form. Display text
stays Cyrillic.

No emitters of Cache-Tag: long remain, making `make clear-long-cache` a
no-op; left in place as a manual escape hatch.

+++ index.js
@@ -324,21 +324,21 @@ app.get("/last-modified.txt", (req, res, next) => {
 app.get("/favicon.ico", (req, res, next) => {
   res
     .set("Cache-Control", `public, max-age=0, s-maxage=${3600 * 24 * 31}`)
-    .set("Cache-Tag", "long");
+    .set("Cache-Tag", "short");
   res.sendFile(path.join(__dirname, "favicon.ico"));
 });
 
 app.get("/smithery.json", (req, res) => {
   res
     .set("Cache-Control", `public, max-age=0, s-maxage=${3600 * 24 * 7}`)
-    .set("Cache-Tag", "long");
+    .set("Cache-Tag", "short");
   res.sendFile(path.join(__dirname, "smithery.json"));
 });
 
 app.get("/server.json", (req, res) => {
   res
     .set("Cache-Control", `public, max-age=0, s-maxage=${3600 * 24 * 7}`)
-    .set("Cache-Tag", "long");
+    .set("Cache-Tag", "short");
   res.sendFile(path.join(__dirname, "server.json"));
 });
