+++ tests/actions/routeStaticInfoAction.test.js
@@ -96,7 +96,7 @@ describe("routeStaticInfoAction", () => {
       "Cache-Control",
       "public, max-age=0, s-maxage=2592000",
     );
-    expect(res.set).toHaveBeenCalledWith("Cache-Tag", "long");
+    expect(res.set).toHaveBeenCalledWith("Cache-Tag", "short");
     expect(res.send).toHaveBeenCalledWith(
       expect.objectContaining({
         id: "EXT1",
